tests: retry createSession to deal with Node registration race condition
Tests would often (10% of the time or so) fail with an error like:

> testCreateSession:               2020/12/13 23:33:40 [ERR] http: Request PUT /v1/session/create, error: Missing node registration from=127.0.0.1:45790

When the agent starts up, it registers the agent's node in the consul catalog.
While the agent may be "ready", the node isn't always registered, and these
session tests would fail if that node was not yet ready. This would happen
because the Session API requires the Node referenced in the API request to be
a node registered in the Consul catalog, so the node must be registered before
using the Session API.