deregisterService uses PUT method instead of GET
Previously, the deregisterService used a GET request to the deregister
endpoint. Consul requires that this must be a PUT request now. This
change converts the request into a PUT request using the createRequest
convention that empty string request bodies are PUT requests.

This change also adds a test for deregisterService. A service is
created, then destroyed, and confirmed that the service was removed.

This change also fixes a minor bug in the testRegisterService. It did
not check that the list returned wasn't empty. If the list of services
is empty, it also wasn't created.