test/Util: update consulServerSetupFunc to ping /v1/status/leader API before running
While we've given consul a bunch of ports, and while we've waited on
each of those to be sure consul is listening, consul server may not yet
be 100% ready (even if the port is open, raft consensus/leadership maybe
pending. By pinging the /v1/status/leader API, we can delay until the
local consul server says it has acquired leadership and is ready for its
consul duties. This ensures consul is ready for tests, and that we don't
fail tests simply because consul isn't 100% ready for them.

However, we don't actually do anything with the response just yet. This
is the plumbing we need to interact with the API and take decisive
action on the response. In another update, we'll handle the loop/retry
portion of this.