syd-test/Utils: add and use functions to block while we wait for consul to aquire leadership
This is a really important update. Prior to this, the tests would spin
up consul and then immediately run the tests (accessing Consul).
Sometimes the tests would try to interact with Consul before the agent
had acquired and settled leadership.

With this update, we ensure that the tests are delayed for a small bit
of time until the Consul agent has acquired leadership.