test/Util: add consulServerSetupFuncWith and include support for TestsuiteSettings using a defaultTestsuiteSettings
add a new data type for the test suite's settings
(enable/disable verbose logs, debug mode, acls, etc)

In particular, focus on settings we need to use right now,
where "need" is based on what we've been flipping comments
on, and tests we're actively developing (ACLs).

Note that we replace the previous `consulServerSetupFunc` with
a new `consulServerSetupFuncWith`. The new `consulServerSetupFunc`
is now a wrapper that calls `consulServerSetupFuncWith` and
provides default settings. In that way, all tests currently
using this function can continue to do so with no change, and
they automatically get the updates we're making.

This gives tests the "default settings" by default, and only if
a test sets up its own settings and call `consulServerSetupFuncWith`
directly will it get non-default settings.

In this context, the "settings" are "details regarding how we
start and configure consul to run.