test/Util: define consul config file contents, acls enabled or not?
If acls are enabled, we need a bunch of attributes in the HCL config.
If acls are not enabled, we have a pretty simple default config.

One problem is that, if acls are enabled, we cannot set a default policy
to deny, as bootstrap acls then fails. Thus, at least for now, we set
our default policy to allow, and bootstrap acls passes :)